add id to country entity and refactor DAO
